backend/requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
numpy
pydantic
sentence-transformers
llama-cpp-python
google-generativeai